refactor ('git-push-tag' action): merge conditional steps back into one and use string comparison (refactor/push-tag-action-merge) [#47]
reason: GitHub Service Bug: booleans are not actual booleans, but strings with either 'true' or 'false' as value.

in depth: ${{ variable }} creates a string with the contents of the variable and does not resolve into a boolean value.
 **This is different from the documentation!**
 Hence we solve this persistent bug by comparing the stringified boolean ${{ inputs.dry-run }}
 with the string "true" to obtain an actual boolean we can use in Powershell.

for the record: The same issue has been raised to GitHub several times.
 e.g. https://github.com/actions/runner/issues/1483

reason: GitHub Service Bug: booleans are not actual booleans, but strings with either 'true' or 'false' as value.

in depth: ${{ variable }} creates a string with the contents of the variable and does not resolve into a boolean value.
 **This is different from the documentation!**
 Hence we solve this persistent bug by comparing the stringified boolean ${{ inputs.dry-run }}
 with the string "true" to obtain an actual boolean we can use in Powershell.

for the record: The same issue has been raised to GitHub several times.
 e.g. https://github.com/actions/runner/issues/1483